requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
duckdb==0.8.0
exceptiongroup==1.1.1
fastapi==0.95.2
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lz4==4.3.2
MarkupSafe==2.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
overrides==7.3.1
packaging==23.1
pandas==2.0.1
posthog==3.0.1
protobuf==4.23.2
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
tqdm==4.65.0
typing-extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
